New Development: No             JIRA Issue:         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): asdm2MS            Description: fix type ambiguity again
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@38250 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/alma/apps/asdm2MS/bdflags2MS.cc
@@ -1191,7 +1191,7 @@ void processCorrelatorFlagsPerSlices(MainRow*					mR_p,
 
   for (unsigned int j = 0; j < actualSizeInMemory.size(); j++) {
     if (debug) cout << "PLATOON" << endl;
-    numberOfIntegrations = min(actualSizeInMemory[j] / (bdfSize / N), N); // The min to prevent a possible excess when there are very few bytes in the BDF.
+    numberOfIntegrations = min((uint64_t)actualSizeInMemory[j] / (bdfSize / N), (uint64_t)N); // The min to prevent a possible excess when there are very few bytes in the BDF.
     if (debug) cout << "------------> " << numberOfIntegrations << " , " << actualSizeInMemory[j] << endl;
 
     if (numberOfIntegrations) {
